Replace remaining static ".UK" references.
Good grief…  Whoever came up with this codebase really should be subjected to some form of torture outlawed by Geneva conventions.  All this, and I'm dead sure when I test it that I'm not gonna get anywhere with it.  But, as long as the silly bugger keeps knocking me around about this, I'll at least try to look like I'm making headway. :-(